refactor: file structure for top navigation bar
