+++ lib/app/view/app.dart
@@ -1,6 +1,6 @@
 import 'package:ax_dapp/app/bloc/app_bloc.dart';
 import 'package:ax_dapp/app/view/app_routing.dart';
-import 'package:ax_dapp/app/widgets/bloc/top_navigation_bar_bloc.dart';
+import 'package:ax_dapp/app/widgets/top_navigation_bar/bloc/top_navigation_bar_bloc.dart';
 import 'package:ax_dapp/league/league_search/bloc/league_bloc.dart';
 import 'package:ax_dapp/league/repository/prize_pool_repository.dart';
 import 'package:ax_dapp/league/usecases/league_use_case.dart';
