+++ lib/app/widgets/widgets.dart
@@ -1,4 +1,3 @@
 export 'bottom_navigation_bar_mobile.dart';
 export 'bottom_navigation_bar_web.dart';
-export 'top_navigation_bar_mobile.dart';
-export 'top_navigation_bar_web.dart';
+export 'top_navigation_bar/top_navigation_bar.dart';
